refactor: solution projects
add projects in order to make the
tests easier

resolves #18
