+++ IssTracker.Application.Services/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace IssTracker.Application.Services
+{
+    public class Class1
+    {
+    }
+}
