+++ IssTracker.Application/Class1.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace IssTracker.Application
+{
+    public class Class1
+    {
+    }
+}
