+++ IssTracker.Presentation.Api/IssTracker.Presentation.Api.csproj
@@ -10,6 +10,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\IssTracker.App.Messages\IssTracker.App.Messages.csproj" />
     <ProjectReference Include="..\IssTracker.Infra.CrossCutting.DepencyInversion\IssTracker.Infra.CrossCutting.DepencyInversion.csproj" />
   </ItemGroup>
