+++ IssTracker.Presentation.Api/appsettings.Development.json
@@ -5,5 +5,15 @@
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     }
+  },
+  "AllowedHosts": "*",
+  "MapApiConfiguration": {
+    "URL": "https://api.opencagedata.com",
+    "URI": "/geocode/v1/json",
+    "KEY": "70e8048eed014bc78cce9e9565a008d4"
+  },
+  "TrackApiConfiguration": {
+    "URL": "http://api.open-notify.org",
+    "URI": "/iss-now.json"
   }
 }
